@@ -1,32 +1,32 @@
 name: linkcheck
-version: 2.0.23 # Don't forget to update in lib/linkcheck.dart, too.
+version: 3.0.0-dev # Don't forget to update in lib/linkcheck.dart, too.
 
 description: >-
   A very fast link-checker. Crawls sites and checks integrity of links
   both in HTML and in CSS.
 
-homepage: https://github.com/filiph/linkcheck
+repository: https://github.com/filiph/linkcheck
 
 environment:
-  sdk: '>=2.11.99 <3.0.0'
+  sdk: '>=2.18.0 <4.0.0'
 
 dependencies:
-  args: '>=1.5.1 <3.0.0'
+  args: ^2.2.0
   console: ^4.1.0
   csslib: ^0.17.0
-  glob: ^2.0.1
+  glob: ^2.1.0
   html: ^0.15.0
-  logging: ^1.0.1
-  path: ^1.6.2
-  source_span: ^1.5.5
-  stream_channel: ^2.0.0
+  meta: ^1.8.0
+  path: ^1.8.0
+  source_span: ^1.8.0
+  stream_channel: ^2.1.0
 
 dev_dependencies:
+  lints: ^2.0.1
   dhttpd: ^4.0.0
-  pedantic: ^1.7.0
-  test: ^1.6.3
-  cli_pkg: ^2.1.1
-  grinder: ^0.9.0
+  test: ^1.22.1
+  cli_pkg: ^2.1.6
+  grinder: ^0.9.2
 
 executables:
   linkcheck: linkcheck
