@@ -1,12 +1,10 @@
-library linkcheck.link;
-
-import 'package:linkcheck/src/destination.dart';
-import 'package:linkcheck/src/origin.dart';
+import 'destination.dart';
+import 'origin.dart';
 
 class Link {
-  Origin origin;
+  final Origin origin;
   Destination destination;
-  String fragment;
+  final String? fragment;
 
   /// Whether or not this link was marked as skipped.
   ///
@@ -14,16 +12,9 @@ class Link {
   /// has a match, [wasSkipped] will be `true`.
   bool wasSkipped = false;
 
-  Link(this.origin, this.destination, String fragment,
+  Link(this.origin, this.destination, String? fragment,
       [this.wasSkipped = false])
       : fragment = fragment == null || fragment.isEmpty ? null : fragment;
-
-  Link.fromMap(Map<String, Object> map)
-      : this(
-            Origin.fromMap(map["origin"] as Map<String, Object>),
-            Destination.fromMap(map["destination"] as Map<String, Object>),
-            map["destinationAnchor"] as String,
-            map["wasSkipped"] as bool);
 
   bool get breaksAnchor =>
       !wasSkipped &&
@@ -56,15 +47,8 @@ class Link {
   bool hasWarning(bool shouldCheckAnchors) =>
       (shouldCheckAnchors && breaksAnchor) || destination.hasNoMimeType;
 
-  Map<String, Object> toMap() => {
-        "origin": origin.toMap(),
-        "destination": destination.toMap(),
-        "destinationAnchor": fragment,
-        "wasSkipped": wasSkipped
-      };
-
   @override
   String toString() => "$origin => $destination"
-      "${fragment == null ? '' : '#' + fragment} "
+      "${fragment == null ? '' : '#$fragment'} "
       "(${destination.statusDescription})";
 }
