@@ -1,4 +1,4 @@
-library linkcheck.server_info;
+import 'package:meta/meta.dart';
 
 import 'parsers/robots_txt.dart';
 
@@ -11,11 +11,11 @@ class ServerInfo {
   /// No duration.
   static const immediate = Duration();
 
-  String host;
+  final String host;
 
-  int port;
+  final int? port;
 
-  RobotsBouncer bouncer;
+  RobotsBouncer? bouncer;
 
   /// The total count of connection attempts, both successful and failed.
   int connectionAttempts = 0;
@@ -39,7 +39,7 @@ class ServerInfo {
   /// or 405).
   bool hasFailedHeadRequest = false;
 
-  DateTime _lastRequest;
+  DateTime? _lastRequest;
 
   ServerInfo(String authority)
       : host = authority.split(':').first,
@@ -57,9 +57,10 @@ class ServerInfo {
   /// Creates the minimum duration to wait before the server should be bothered
   /// again.
   Duration getThrottlingDuration() {
-    if (_lastRequest == null) return immediate;
+    var lastRequest = _lastRequest;
+    if (lastRequest == null) return immediate;
     if (isLocalhost) return immediate;
-    var sinceLastRequest = DateTime.now().difference(_lastRequest);
+    var sinceLastRequest = DateTime.now().difference(lastRequest);
     if (sinceLastRequest.isNegative) {
       // There's a request scheduled in the future.
       return -sinceLastRequest + minimumDelay;
@@ -83,7 +84,7 @@ class ServerInfo {
         forRobot: robotName);
   }
 
-  void updateFromStatusCode(int statusCode) {
+  void updateFromStatusCode(int? statusCode) {
     connectionAttempts += 1;
     if (statusCode == null) {
       didNotConnectCount += 1;
@@ -107,21 +108,14 @@ class ServerInfo {
 }
 
 /// To be sent from Worker to main thread.
+@immutable
 class ServerInfoUpdate {
-  String host;
-  bool didNotConnect = false;
-  String robotsTxtContents = "";
+  final String host;
+  final bool didNotConnect;
+  final String robotsTxtContents;
 
-  ServerInfoUpdate(this.host);
-  ServerInfoUpdate.fromMap(Map<String, Object> map)
-      : this._(map["host"] as String, map["robots"] as String,
-            map["didNotConnect"] as bool);
+  ServerInfoUpdate(this.host,
+      {this.robotsTxtContents = '', this.didNotConnect = false});
 
-  ServerInfoUpdate._(this.host, this.robotsTxtContents, this.didNotConnect);
-
-  Map<String, Object> toMap() => {
-        "host": host,
-        "robots": robotsTxtContents,
-        "didNotConnect": didNotConnect
-      };
+  ServerInfoUpdate.didNotConnect(String host) : this(host, didNotConnect: true);
 }
