@@ -1,22 +1,14 @@
-library linkcheck.fetch_results;
-
 import '../destination.dart';
 import '../link.dart';
 
 class FetchResults {
   final DestinationResult checked;
   final List<Link> links;
-  FetchResults(this.checked, this.links);
 
-  FetchResults.fromMap(Map<String, Object> map)
-      : this(
-            DestinationResult.fromMap(map["checked"] as Map<String, Object>),
-            List<Link>.from((map["links"] as List<Map>).map<Link>(
-                (serialization) =>
-                    Link.fromMap(serialization as Map<String, Object>))));
+  FetchResults(this.checked, [this.links = const []]);
 
-  Map<String, Object> toMap() => {
-        "checked": checked.toMap(),
-        "links": links?.map((link) => link.toMap())?.toList() ?? <Object>[]
-      };
+  @override
+  String toString() {
+    return 'FetchResults{checked: $checked, links: $links}';
+  }
 }
